rename string format
