+++ golem-ts-agentic/packages/golem-ts-sdk/src/internal/mapping/types/stringFormat.ts
